rename player > Player
