+++ plugins/Player/CMakeLists.txt
