+++ plugins/Player/Player.qml
