+++ plugins/Player/audio/high_hat_1(small).mp3
