+++ plugins/Player/audio/high_hat_1.mp3
