+++ plugins/Player/player.qrc
