+++ plugins/Player/playerplugin.cpp
