+++ plugins/Player/soundthread.cpp
